+++ components/button/presets/0-default.json
@@ -1,0 +1,11 @@
+{
+	"rootId": "button3703996384",
+	"elements": {
+	  "button3703996384": {
+		"name": "Button",
+		"props": {
+		  "label": "Merge"
+		}
+	  }
+	}
+  }
