+++ components/button/presets/0-default.jsx
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import Button from '../button.jsx';
+
+export default <Button label="merge!" />;
